+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.6
 
-- change footer hide show function.
+- change footer hide show function,fix a bug in remove listeners.